requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=1.5.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
